+++ CHANGELOG.md
@@ -4,10 +4,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## unreleased
 
-Feature:
+Breaking changes:
 - Rename the module `is-plain-number` to `is-string-a-number`: after some
   thoughts, I realised that the purpose of my module was to specifically check
   if a `string` value was representing a number in decimal base and notation.
+- Only check `string` values. `number` values are not accepted anymore.
+
 
 ## [v1.0.1] - 2018-11-03